+++ dephell/converters/pip.py
@@ -126,9 +126,10 @@ class PIPConverter(BaseConverter):
         if req.editable:
             line += '-e '
         if req.link is not None:
+            req.link.name = req.name  # patch `#=egg` by right name
             line += req.link.long
         else:
-            line += req.name
+            line += req.raw_name
         if req.extras:
             line += '[{extras}]'.format(extras=','.join(req.extras))
         if req.version: