+++ dephell/converters/setuppy.py
@@ -292,7 +292,7 @@ class SetupPyConverter(BaseConverter):
 
     @staticmethod
     def _format_req(req) -> str:
-        line = req.name
+        line = req.raw_name
         if req.extras:
             line += '[{extras}]'.format(extras=','.join(req.extras))
         if req.version: